refactor(service-sg-server): change import file to shared
https://github.com/akky-xxxx/project-my-homepage/issues/91
