+++ packages/service-sg-server/src/handlers/locationsGet/index.ts
@@ -1,11 +1,11 @@
 import { Locations, Photos } from "common-types"
 import { LocationsGETRes } from "../../libs/bffApiClient"
+import { Query } from "../../shared/const/strapi/Query"
 import { apiHandler } from "../../shared/utils/apiHandler"
 import { extractImagePath } from "../../shared/utils/extractImagePath"
 import { isErrorStatus } from "../../shared/utils/isErrorStatus"
 import { infoLogger, loggerWrapper } from "../../shared/utils/logger"
 import { strapiApiClient } from "../../shared/utils/strapiApiClient"
-import { Query } from "./const"
 import { getClientLocations } from "./modules/getClientLocations"
 
 type LocationsGetBase = () => Promise<LocationsGETRes>
