+++ packages/service-sg-server/src/handlers/locationsIdGet/index.ts
@@ -1,11 +1,11 @@
 import { LocationsId, Photos, Tags } from "common-types"
 import { Request } from "express"
 import { LocationsIdGETRes } from "../../libs/bffApiClient"
+import { Query } from "../../shared/const/strapi/Query"
 import { apiHandler } from "../../shared/utils/apiHandler"
 import { isErrorStatus } from "../../shared/utils/isErrorStatus"
 import { infoLogger, loggerWrapper } from "../../shared/utils/logger"
 import { strapiApiClient } from "../../shared/utils/strapiApiClient"
-import { Query } from "./const"
 import { getClientImage } from "./modules/getClientImage"
 
 type TagsIdGetBase = (
