+++ packages/service-sg-server/src/handlers/photoGalleryGet/index.ts
@@ -1,10 +1,10 @@
 import { Card, Photos, ResponseBody } from "common-types"
 import { PhotoGalleryGETRes } from "../../libs/bffApiClient"
+import { Query } from "../../shared/const/strapi/Query"
 import { apiHandler } from "../../shared/utils/apiHandler"
 import { isErrorStatus } from "../../shared/utils/isErrorStatus"
 import { infoLogger, loggerWrapper } from "../../shared/utils/logger"
 import { strapiApiClient } from "../../shared/utils/strapiApiClient"
-import { Query } from "./const"
 import { getCard } from "./modules/getCard"
 
 type PhotoGalleryGetBase = () => Promise<PhotoGalleryGETRes>
